SANITIZER and SANITIZER_THREAD and fixes

+++ gempyrelib/src/base64.cpp
@@ -26,7 +26,7 @@ std::string Base64::encode(const Byte* buf, size_t bufLen) {
   while (bufLen--) {
     char_array_3[i++] = *(buf++);
     if (i == 3) {
-      char_array_4[0] = (char_array_3[0] & 0xfc) >> 2;
+      char_array_4[0] = static_cast<Byte>((char_array_3[0] & 0xfc) >> 2);
       char_array_4[1] = static_cast<Byte>(((char_array_3[0] & 0x03) << 4) + ((char_array_3[1] & 0xf0) >> 4));
       char_array_4[2] = static_cast<Byte>(((char_array_3[1] & 0x0f) << 2) + ((char_array_3[2] & 0xc0) >> 6));
       char_array_4[3] = char_array_3[2] & 0x3f;
@@ -42,7 +42,7 @@ std::string Base64::encode(const Byte* buf, size_t bufLen) {
     for(unsigned j = i; j < 3; j++)
       char_array_3[j] = '\0';
 
-    char_array_4[0] = (char_array_3[0] & 0xfc) >> 2;
+    char_array_4[0] = static_cast<Byte>((char_array_3[0] & 0xfc) >> 2);
     char_array_4[1] = static_cast<Byte>(((char_array_3[0] & 0x03) << 4) + ((char_array_3[1] & 0xf0) >> 4));
     char_array_4[2] = static_cast<Byte>(((char_array_3[1] & 0x0f) << 2) + ((char_array_3[2] & 0xc0) >> 6));
     char_array_4[3] = char_array_3[2] & 0x3f;
